# Pack `Tile` into 64 bits