--- a/sparse_strips/vello_common/src/tile.rs
+++ b/sparse_strips/vello_common/src/tile.rs
@@ -24,8 +24,8 @@ pub struct Tile(
     ///
     /// - `[ 0, 31)`: The 31-bit index of the line this tile belongs to into the line buffer;
     /// - `[31, 32)`: The 1-bit coarse winding of the tile. This is `1` if and only if the lines
-    /// crosses the tile's top edge. Lines making this crossing increment or decrement the coarse
-    /// tile winding, depending on the line direction;
+    ///   crosses the tile's top edge. Lines making this crossing increment or decrement the coarse
+    ///   tile winding, depending on the line direction;
     /// - `[32, 48)`: The 16-bit x-coordinate; and
     /// - `[48, 64)`: The 16-bit y-coordinate.
     ///