--- a/sparse_strips/vello_common/src/tile.rs
+++ b/sparse_strips/vello_common/src/tile.rs
@@ -22,12 +22,12 @@ pub struct Tile(
     ///
     /// The layout is as follows, with the bit indices from least to most significant.
     ///
-    /// [ 0, 31): The 31-bit index of the line this tile belongs to into the line buffer;
-    /// [31, 32): The 1-bit coarse winding of the tile. This is `1` if and only if the lines
+    /// - `[ 0, 31)`: The 31-bit index of the line this tile belongs to into the line buffer;
+    /// - `[31, 32)`: The 1-bit coarse winding of the tile. This is `1` if and only if the lines
     /// crosses the tile's top edge. Lines making this crossing increment or decrement the coarse
     /// tile winding, depending on the line direction;
-    /// [32, 48): The 16-bit x-coordinate; and
-    /// [48, 64): The 16-bit y-coordinate.
+    /// - `[32, 48)`: The 16-bit x-coordinate; and
+    /// - `[48, 64)`: The 16-bit y-coordinate.
     ///
     /// Note the byte layout in memory depends on the endianness of the compilation target.
     pub u64,