(Split off from https://github.com/linebender/vello/pull/847.)

The fields of `Tile` are now packed such that its value in memory can readily be interpreted as a `u64` consisting, from most to least significant, of the numeric values of 16-bit `y`, 16-bit `x`, 1-bit `winding`, and 31-bit `line_index`. This means its value in memory can be used directly for sorting the tiles.

Sorting on a `u32` of only `y` and `x` is slightly faster for sorting (by about ~0.3 ms on paris-30k), but sorting the entire `u64` speeds up strip rendering (by about ~1.5ms on paris-30k). Presumably sorted line indices help cache locality during strip rendering, and ordering the winding bit before the line index also improves strip rendering performance, maybe because of branch predictability during strip rendering (see [this comment](https://github.com/linebender/vello/pull/847#discussion_r1990081159)).

On my machine, averaging over 500 iterations paris-30k on main and in this PR measure as:

```
Tile generation
    before: 17.329454ms
    after:  16.633348ms
Tile sorting
    before: 8.120621ms
    after:  7.3823915ms
Strip generation
    before: 31.896929ms
    after:  29.898085ms
```