@LaurenzV 2025-03-13
> The order of components from most to least significant is 16-bit y and x, 1-bit winding, and 31-bit line_index. Sorting on a u32 of only y and x is slightly faster for sorting (by about ~0.3 ms on paris-30k), but sorting the entire u64 speeds up strip rendering (by about ~1.5ms on paris-30k). 

Have you also tried u32 but with everything packed, i.e. the original suggestion by Raph? Would at least be interesting to try to see whether there are any performance gains, or whether just sticking to u64 is the best.