@tomcur 2025-03-13
> Have you also tried u32 but with everything packed, i.e. the original suggestion by Raph? Would at least be interesting to try to see whether there are any performance gains, or whether just sticking to u64 is the best.

Not yet, but I agree it would be interesting to test.