@tomcur 2025-03-13
See https://godbolt.org/z/44ffv3v3G for codegen examples depending on field order.